remove_pbc.cpp
This set of functions should enable us to get the distance between two molecules while ignoring the periodic boundary conditions, after I update the main.cpp. It may need some more work. For example, if a copper is at the bottom of the box and an ammonia is at the top of the box their distance will be calculated from bottom to top so it will appear very large, but with this function it will take the absolute distance from the ammonia at the top and go over the pbc to the copper at the bottom and get a much smaller distance.